refactor: 패키지 이동 및 사용 안하는 파일 제거
