--- android/app/src/main/java/com/on/turip/ui/common/extensions/DpFormatExtensions.kt
@@ -1,9 +1,0 @@
-package com.on.turip.ui.common.extensions
-
-import android.content.res.Resources
-
-val Int.dpToPxFloat: Float
-    get() = (this * Resources.getSystem().displayMetrics.density + 0.5f)
-
-val Int.dpToPx: Int
-    get() = (this * Resources.getSystem().displayMetrics.density + 0.5f).toInt()
