+++ android/app/src/main/java/com/on/turip/ui/common/mapper/TripMapper.kt
@@ -6,8 +6,8 @@ import com.on.turip.domain.trip.Trip
 import com.on.turip.domain.trip.TripDuration
 import com.on.turip.ui.common.model.trip.TripDurationModel
 import com.on.turip.ui.common.model.trip.TripModel
+import com.on.turip.ui.compose.home.model.UsersLikeContentModel
 import com.on.turip.ui.compose.trip.model.PlaceModel
-import com.on.turip.ui.main.home.model.UsersLikeContentModel
 
 fun Trip.toUiModel(): TripModel =
     TripModel(
