+++ android/app/src/main/java/com/on/turip/ui/common/mapper/TuripMapper.kt
@@ -1,22 +1,43 @@
 package com.on.turip.ui.common.mapper
 
+import com.google.android.gms.maps.model.LatLng
+import com.on.turip.domain.bookmark.TuripPlace
 import com.on.turip.domain.turip.Turip
-import com.on.turip.ui.folder.model.TuripEditModel
-import com.on.turip.ui.main.favorite.model.TuripModel
+import com.on.turip.ui.common.model.turip.TuripModel
+import com.on.turip.ui.common.model.turip.TuripPlaceShareModel
+import com.on.turip.ui.compose.trip.turipselection.model.TuripPlaceModel
+import com.on.turip.ui.compose.turipdetail.model.turip.PlaceLatLngUiModel
 
-fun Turip.toUiModel(selectTuripId: Long): TuripModel =
+fun TuripPlace.toUiModel(): TuripPlaceModel =
+    TuripPlaceModel(
+        turipPlaceId = id,
+        placeId = place.placeId,
+        order = order,
+        name = place.name,
+        isTuripPlace = true,
+        latLng = LatLng(place.latitude, place.longitude),
+        category = place.category.joinToString(),
+        mapLink = place.url,
+    )
+
+fun TuripPlace.toLatLng(): PlaceLatLngUiModel =
+    PlaceLatLngUiModel(
+        placeId = place.placeId,
+        name = place.name,
+        latLng = LatLng(place.latitude, place.longitude),
+    )
+
+fun Turip.toUiModel(): TuripModel =
     TuripModel(
         id = id,
         name = name,
         placeCount = placeCount,
-        isSelected = id == selectTuripId,
+        isSelected = hasIncludePlace,
         isDefault = isDefault,
     )
 
-fun Turip.toEditUiModel(): TuripEditModel =
-    TuripEditModel(
-        id = id,
+fun TuripPlaceModel.toUiModel(): TuripPlaceShareModel =
+    TuripPlaceShareModel(
         name = name,
-        count = placeCount,
-        isSelected = false,
+        uri = mapModel.uri,
     )
