+++ android/app/src/main/java/com/on/turip/ui/common/model/turip/TuripPlaceShareModel.kt
@@ -1,4 +1,4 @@
-package com.on.turip.ui.main.favorite.model
+package com.on.turip.ui.common.model.turip
 
 import android.net.Uri
