+++ android/app/src/main/java/com/on/turip/ui/compose/home/HomeScreen.kt
@@ -42,7 +42,7 @@ import com.on.turip.ui.compose.home.component.RegionList
 import com.on.turip.ui.compose.home.component.RegionTypeButtons
 import com.on.turip.ui.compose.home.component.SearchTextField
 import com.on.turip.ui.compose.home.component.UsersLikeList
-import com.on.turip.ui.main.home.model.UsersLikeContentModel
+import com.on.turip.ui.compose.home.model.UsersLikeContentModel
 import kotlinx.coroutines.flow.collectLatest
 
 @Composable
