+++ android/app/src/main/java/com/on/turip/ui/compose/home/model/UsersLikeContentModel.kt
@@ -1,4 +1,4 @@
-package com.on.turip.ui.main.home.model
+package com.on.turip.ui.compose.home.model
 
 import com.on.turip.domain.content.Content
 import com.on.turip.ui.common.model.trip.TripDurationModel
