+++ android/app/src/main/java/com/on/turip/ui/compose/trip/navigation/TripDetailNavKeyProvider.kt
@@ -9,8 +9,8 @@ import com.on.turip.R
 import com.on.turip.navigation.NavKeyProvider
 import com.on.turip.navigation.Navigator
 import com.on.turip.ui.common.extensions.safeStartActivityWithToast
+import com.on.turip.ui.common.model.turip.TuripShareModel
 import com.on.turip.ui.compose.trip.model.MapModel
-import com.on.turip.ui.compose.turipdetail.model.turip.TuripShareModel
 import dagger.hilt.android.qualifiers.ActivityContext
 import kotlinx.serialization.modules.PolymorphicModuleBuilder
 import javax.inject.Inject
