+++ android/app/src/main/java/com/on/turip/ui/compose/trip/turipselection/PlaceTuripSelectionBottomSheet.kt
@@ -30,6 +30,7 @@ import androidx.lifecycle.repeatOnLifecycle
 import com.on.turip.R
 import com.on.turip.ui.common.error.ErrorUiModel
 import com.on.turip.ui.common.error.toUiModel
+import com.on.turip.ui.common.model.turip.TuripShareModel
 import com.on.turip.ui.compose.designsystem.component.TuripDialog
 import com.on.turip.ui.compose.designsystem.component.TuripSnackbar
 import com.on.turip.ui.compose.designsystem.snackbar.LocalSnackbarDelegate
@@ -39,7 +40,6 @@ import com.on.turip.ui.compose.trip.model.SelectedPlaceModel
 import com.on.turip.ui.compose.trip.turipselection.component.PlaceTuripSelectionContent
 import com.on.turip.ui.compose.trip.turipselection.component.ShareOptionBottomSheet
 import com.on.turip.ui.compose.trip.turipselection.component.TuripsContent
-import com.on.turip.ui.compose.turipdetail.model.turip.TuripShareModel
 import kotlinx.collections.immutable.persistentListOf
 
 @OptIn(ExperimentalMaterial3Api::class)
