+++ android/app/src/main/java/com/on/turip/ui/compose/trip/turipselection/PlaceTuripSelectionUiState.kt
@@ -1,8 +1,8 @@
 package com.on.turip.ui.compose.trip.turipselection
 
 import androidx.compose.runtime.Immutable
+import com.on.turip.ui.common.model.turip.TuripModel
 import com.on.turip.ui.compose.trip.turipselection.model.TuripPlaceModel
-import com.on.turip.ui.main.favorite.model.TuripModel
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
