+++ android/app/src/main/java/com/on/turip/ui/compose/trip/turipselection/component/TuripItem.kt
@@ -23,8 +23,8 @@ import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import androidx.compose.ui.unit.dp
 import com.on.turip.R
+import com.on.turip.ui.common.model.turip.TuripModel
 import com.on.turip.ui.compose.designsystem.theme.TuripTheme
-import com.on.turip.ui.main.favorite.model.TuripModel
 
 @Composable
 fun TuripItem(
